requirements.txt
gymnasium
pygame
numpy 
pandas
matplotlib
scikit-learn
pytorch
